@@ -1,28 +1,31 @@
-library fancy_bottom_navigation;
+library fancy_bottom_navigation_image;
 
-import 'package:fancy_bottom_navigation/internal/tab_item.dart';
-import 'package:fancy_bottom_navigation/paint/half_clipper.dart';
-import 'package:fancy_bottom_navigation/paint/half_painter.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-const double CIRCLE_SIZE = 60;
+import 'internal/tab_item.dart';
+import 'paint/half_clipper.dart';
+import 'paint/half_painter.dart';
+
+
+const double CIRCLE_SIZE = 50;
 const double ARC_HEIGHT = 70;
 const double ARC_WIDTH = 90;
-const double CIRCLE_OUTLINE = 10;
-const double SHADOW_ALLOWANCE = 20;
+const double CIRCLE_OUTLINE = 5;
+const double SHADOW_ALLOWANCE = 30;
 const double BAR_HEIGHT = 60;
 
-class FancyBottomNavigation extends StatefulWidget {
-  FancyBottomNavigation(
+class FancyBottomNavigationImage extends StatefulWidget {
+  FancyBottomNavigationImage(
       {@required this.tabs,
-      @required this.onTabChangedListener,
-      this.key,
-      this.initialSelection = 0,
-      this.circleColor,
-      this.activeIconColor,
-      this.inactiveIconColor,
-      this.textColor,
-      this.barBackgroundColor})
+        @required this.onTabChangedListener,
+        this.key,
+        this.initialSelection = 0,
+        this.circleColor,
+        this.activeIconColor,
+        this.inactiveIconColor,
+        this.textColor,
+        this.barBackgroundColor})
       : assert(onTabChangedListener != null),
         assert(tabs != null),
         assert(tabs.length > 1 && tabs.length < 5);
@@ -39,13 +42,13 @@ class FancyBottomNavigation extends StatefulWidget {
   final Key key;
 
   @override
-  FancyBottomNavigationState createState() => FancyBottomNavigationState();
+  FancyBottomNavigationImageState createState() => FancyBottomNavigationImageState();
 }
 
-class FancyBottomNavigationState extends State<FancyBottomNavigation>
+class FancyBottomNavigationImageState extends State<FancyBottomNavigationImage>
     with TickerProviderStateMixin, RouteAware {
-  IconData nextIcon = Icons.search;
-  IconData activeIcon = Icons.search;
+  String nextIcon;
+  String activeIcon;
 
   int currentSelected = 0;
   double _circleAlignX = 0;
@@ -61,34 +64,34 @@ class FancyBottomNavigationState extends State<FancyBottomNavigation>
   void didChangeDependencies() {
     super.didChangeDependencies();
 
-    activeIcon = widget.tabs[currentSelected].iconData;
+    activeIcon = widget.tabs[currentSelected].imageData;
 
     circleColor = (widget.circleColor == null)
         ? (Theme.of(context).brightness == Brightness.dark)
-            ? Colors.white
-            : Theme.of(context).primaryColor
+        ? Colors.white
+        : circleColor
         : widget.circleColor;
 
     activeIconColor = (widget.activeIconColor == null)
         ? (Theme.of(context).brightness == Brightness.dark)
-            ? Colors.black54
-            : Colors.white
+        ? Colors.black54
+        : Colors.white
         : widget.activeIconColor;
 
     barBackgroundColor = (widget.barBackgroundColor == null)
         ? (Theme.of(context).brightness == Brightness.dark)
-            ? Color(0xFF212121)
-            : Colors.white
+        ? Color(0xFF212121)
+        : Colors.white
         : widget.barBackgroundColor;
     textColor = (widget.textColor == null)
         ? (Theme.of(context).brightness == Brightness.dark)
-            ? Colors.white
-            : Colors.black54
+        ? Colors.white
+        : Colors.black54
         : widget.textColor;
     inactiveIconColor = (widget.inactiveIconColor == null)
         ? (Theme.of(context).brightness == Brightness.dark)
-            ? Colors.white
-            : Theme.of(context).primaryColor
+        ? Colors.white
+        : inactiveIconColor
         : widget.inactiveIconColor;
   }
 
@@ -105,7 +108,7 @@ class FancyBottomNavigationState extends State<FancyBottomNavigation>
       setState(() {
         currentSelected = selected;
         _circleAlignX = -1 + (2 / (widget.tabs.length - 1) * selected);
-        nextIcon = widget.tabs[selected].iconData;
+        nextIcon = widget.tabs[selected].imageData;
       });
     }
   }
@@ -127,19 +130,19 @@ class FancyBottomNavigationState extends State<FancyBottomNavigation>
             crossAxisAlignment: CrossAxisAlignment.center,
             children: widget.tabs
                 .map((t) => TabItem(
-                    uniqueKey: t.key,
-                    selected: t.key == widget.tabs[currentSelected].key,
-                    iconData: t.iconData,
-                    title: t.title,
-                    iconColor: inactiveIconColor,
-                    textColor: textColor,
-                    callbackFunction: (uniqueKey) {
-                      int selected = widget.tabs
-                          .indexWhere((tabData) => tabData.key == uniqueKey);
-                      widget.onTabChangedListener(selected);
-                      _setSelected(uniqueKey);
-                      _initAnimationAndStart(_circleAlignX, 1);
-                    }))
+                uniqueKey: t.key,
+                selected: t.key == widget.tabs[currentSelected].key,
+                imageData: t.imageData,
+                title: t.title,
+                iconColor: inactiveIconColor,
+                textColor: textColor,
+                callbackFunction: (uniqueKey) {
+                  int selected = widget.tabs
+                      .indexWhere((tabData) => tabData.key == uniqueKey);
+                  widget.onTabChangedListener(selected);
+                  _setSelected(uniqueKey);
+                  _initAnimationAndStart(_circleAlignX, 1);
+                }))
                 .toList(),
           ),
         ),
@@ -161,33 +164,36 @@ class FancyBottomNavigationState extends State<FancyBottomNavigation>
                       children: <Widget>[
                         SizedBox(
                           height:
-                              CIRCLE_SIZE + CIRCLE_OUTLINE + SHADOW_ALLOWANCE,
+                          CIRCLE_SIZE + CIRCLE_OUTLINE + SHADOW_ALLOWANCE,
                           width:
-                              CIRCLE_SIZE + CIRCLE_OUTLINE + SHADOW_ALLOWANCE,
+                          CIRCLE_SIZE + CIRCLE_OUTLINE + SHADOW_ALLOWANCE,
                           child: ClipRect(
-                              clipper: HalfClipper(),
-                              child: Container(
-                                child: Center(
-                                  child: Container(
-                                      width: CIRCLE_SIZE + CIRCLE_OUTLINE,
-                                      height: CIRCLE_SIZE + CIRCLE_OUTLINE,
-                                      decoration: BoxDecoration(
-                                          color: Colors.white,
-                                          shape: BoxShape.circle,
-                                          boxShadow: [
-                                            BoxShadow(
-                                                color: Colors.black12,
-                                                blurRadius: 8)
-                                          ])),
+                            clipper: HalfClipper(),
+                            child: Container(
+                              child: Center(
+                                child: Container(
+                                  width: CIRCLE_SIZE + CIRCLE_OUTLINE,
+                                  height: CIRCLE_SIZE + CIRCLE_OUTLINE,
+                                  decoration: BoxDecoration(
+                                    color: Colors.white,
+                                    shape: BoxShape.circle,
+                                    boxShadow: [
+                                      BoxShadow(
+                                          color: Colors.black12, blurRadius: 8)
+                                    ],
+                                  ),
                                 ),
-                              )),
+                              ),
+                            ),
+                          ),
                         ),
                         SizedBox(
-                            height: ARC_HEIGHT,
-                            width: ARC_WIDTH,
-                            child: CustomPaint(
-                              painter: HalfPainter(barBackgroundColor),
-                            )),
+                          height: ARC_HEIGHT,
+                          width: ARC_WIDTH,
+                          child: CustomPaint(
+                            painter: HalfPainter(barBackgroundColor),
+                          ),
+                        ),
                         SizedBox(
                           height: CIRCLE_SIZE,
                           width: CIRCLE_SIZE,
@@ -195,12 +201,12 @@ class FancyBottomNavigationState extends State<FancyBottomNavigation>
                             decoration: BoxDecoration(
                                 shape: BoxShape.circle, color: circleColor),
                             child: Padding(
-                              padding: const EdgeInsets.all(0.0),
+                              padding: const EdgeInsets.all(15),
                               child: AnimatedOpacity(
                                 duration:
-                                    Duration(milliseconds: ANIM_DURATION ~/ 5),
+                                Duration(milliseconds: ANIM_DURATION ~/ 5),
                                 opacity: _circleIconAlpha,
-                                child: Icon(
+                                child: Image.asset(
                                   activeIcon,
                                   color: activeIconColor,
                                 ),
@@ -248,9 +254,9 @@ class FancyBottomNavigationState extends State<FancyBottomNavigation>
 }
 
 class TabData {
-  TabData({@required this.iconData, @required this.title, this.onclick});
+  TabData({@required this.imageData, @required this.title, this.onclick});
 
-  IconData iconData;
+  String imageData;
   String title;
   Function onclick;
   final UniqueKey key = UniqueKey();
