@@ -9,8 +9,12 @@ class _SecondPageState extends State<SecondPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: Text("Second page"),),
-      body: Center(child: Text("Nothing to see here"),),
+      appBar: AppBar(
+        title: Text("Second page"),
+      ),
+      body: Center(
+        child: Text("Nothing to see here"),
+      ),
     );
   }
 }
